% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{ba_plot}
\alias{ba_plot}
\title{Generate a before-after plot}
\usage{
ba_plot(DE_before, DE_after, factor_name = "sample_type")
}
\arguments{
\item{DE_before}{The dataset before the process or treatment.}

\item{DE_after}{The dataset after the process or treatment.}

\item{factor_name}{The name of the factor variable for stratification.}
}
\value{
A ggplot object displaying the before-after plot.
}
\description{
This function generates a before-after plot comparing distributions before and after a certain process or treatment.
}
\examples{
ba_plot(DE_before = before_data, DE_after = after_data, factor_name = "sample_type")
}
