% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{batch_plot}
\alias{batch_plot}
\title{Generate a batch plot}
\usage{
batch_plot(
  dataset_experiment,
  order_col,
  batch_col,
  qc_col,
  qc_label,
  colour_by_col,
  feature_to_plot,
  ylab = "Peak area",
  title = "Feature vs run_order"
)
}
\arguments{
\item{dataset_experiment}{The dataset for which the batch plot will be generated.}

\item{order_col}{The column representing run order.}

\item{batch_col}{The column representing batches.}

\item{qc_col}{The column representing QC samples.}

\item{qc_label}{The label for QC samples.}

\item{colour_by_col}{The column used for coloring in the plot.}

\item{feature_to_plot}{The feature to be plotted.}

\item{ylab}{The label for the y-axis.}

\item{title}{The title of the plot.}
}
\value{
A ggplot object displaying the batch plot.
}
\description{
This function generates a batch plot showing the relationship between a feature and run order, stratified by batches and quality control (QC) samples.
}
\examples{
batch_plot(dataset_experiment = my_dataset, order_col = "order", batch_col = "batch", qc_col = "qc", qc_label = "qc_label", colour_by_col = "colour", feature_to_plot = "feature", ylab = "Peak area", title = "Feature vs run_order")
}
