% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{combat_correction}
\alias{combat_correction}
\title{Correct Batch Effects in Metabolomics Data Using ComBat}
\usage{
combat_correction(dataset_exp, batch_col)
}
\arguments{
\item{dataset_exp}{A list containing two elements:
\describe{
\item{data}{A data frame or matrix of metabolite intensities with samples as rows and metabolites as columns.}
\item{sample_meta}{A data frame containing sample metadata, with rows corresponding to samples and a column for batch information.}
}}

\item{batch_col}{A string specifying the column name in \code{sample_meta} that contains the batch information.}
}
\value{
A matrix of corrected metabolite intensities with batch effects removed.
}
\description{
This function applies the ComBat method to correct for batch effects in a metabolomics dataset.
}
\examples{
# Example usage:
DE <- metaboPipe::MTBLS79
corrected_data <- combat_correction(DE, "Batch")

}
