% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data.modify}
\alias{data.modify}
\title{Function to modify data matrix of a DatasetExperiment object}
\usage{
data.modify(dataset_exp, data)
}
\arguments{
\item{dataset_exp}{A DatasetExperiment object.}

\item{data}{New data matrix.}
}
\value{
A DatasetExperiment object with modified data matrix.
}
\description{
This function replaces the data matrix in a SummarizedExperiment object with new data.
}
\examples{
data.modify(dataset_exp, data)
}
