% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{distribution_boxplot}
\alias{distribution_boxplot}
\title{Generate a distribution boxplot}
\usage{
distribution_boxplot(dataset_experiment, factor_name, per_class = FALSE)
}
\arguments{
\item{dataset_experiment}{The dataset for which the boxplot will be generated.}

\item{factor_name}{The name of the factor variable.}

\item{per_class}{Logical indicating whether to generate separate boxplots for each class of the factor.}
}
\value{
A ggplot object displaying the distribution boxplot.
}
\description{
This function generates a distribution boxplot for a specified factor in the dataset.
}
\examples{
distribution_boxplot(dataset_experiment = my_dataset, factor_name = "factor", per_class = TRUE)
}
