% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_factories.R
\name{export_data}
\alias{export_data}
\title{Export Data}
\usage{
export_data(output_name, input_name, out_dir, out_name = "Processed")
}
\arguments{
\item{output_name}{The name of the exported dataset.}

\item{input_name}{The name of the input dataset to be exported.}

\item{out_dir}{The directory where the exported dataset will be saved.}

\item{out_name}{The name of the exported dataset (default is "Processed").}
}
\value{
Nothing, but saves the dataset to the specified directory.
}
\description{
Exports a dataset to a specified directory with a given name.
}
\examples{
export_data("exported_dataset.csv", my_dataset, "output_directory/")
}
