% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_factories.R
\name{factorize_cols}
\alias{factorize_cols}
\title{Factorize columns in sample metadata}
\usage{
factorize_cols(output_name, input_name, cols)
}
\arguments{
\item{output_name}{The name of the output target.}

\item{input_name}{The name of the input data.}

\item{cols}{The columns to factorize.}
}
\value{
A target to factorize columns in sample metadata.
}
\description{
Factorize columns in sample metadata
}
\examples{
factorize_cols(factorized_experiment, input_name = experiment, cols = c("Col1", "Col2"))
}
\seealso{
\code{\link[=warper_factor_sample_col]{warper_factor_sample_col()}}
}
