% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_blanks}
\alias{filter_blanks}
\title{Filter Blanks}
\usage{
filter_blanks(
  dataset_experiment,
  fold_change = 20,
  blank_label = "blank",
  qc_label = "QC",
  factor_name = "sample_type",
  fraction_in_blank = 0
)
}
\arguments{
\item{dataset_experiment}{The \code{DatasetExperiment} object.}

\item{fold_change}{The fold change threshold for blank filtering.}

\item{blank_label}{The label for blanks.}

\item{qc_label}{The label for quality control samples.}

\item{factor_name}{The factor column name.}

\item{fraction_in_blank}{The fraction of values in blank (default is 0).}
}
\value{
A \code{DatasetExperiment} object with blanks filtered out.
}
\description{
Filter blanks from the \code{DatasetExperiment}.
}
\examples{
DE <- metaboPipe::ST000284
filtered_data <- filter_blanks(DE, fold_change = 20, blank_label = "blank", qc_label = "QC", factor_name = "sample_type", fraction_in_blank = 0)
}
