% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_factories.R
\name{load_data}
\alias{load_data}
\title{Load data from files into data frames}
\usage{
load_data(
  output_name,
  dataMatrixFile,
  sampleMetadataFile,
  variableMetadataFile = NULL,
  dataSep = ",",
  sampleSep = ",",
  variableSep = ","
)
}
\arguments{
\item{output_name}{The name of the output target.}

\item{dataMatrixFile}{Path to the data matrix file.}

\item{sampleMetadataFile}{Path to the sample metadata file.}

\item{variableMetadataFile}{Path to the variable metadata file (optional).}

\item{dataSep}{The separator used in the dataMatrixfile (default is ",").}

\item{sampleSep}{The separator used in the sampleMetadataFile (default is ",").}

\item{variableSep}{The separator used in the variableMetadataFile (default is ",").}
}
\value{
A list of targets to load and read data matrix and sample metadata.
}
\description{
Load data from files into data frames
}
\examples{
load_data(data_loaded, "Data/data.csv", "Data/metadata.csv", "Data/variable_metadata.csv", separator = ",")
}
