% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{missing_values_plot}
\alias{missing_values_plot}
\title{Generate a missing values plot}
\usage{
missing_values_plot(dataset_experiment, out_dir, out_name)
}
\arguments{
\item{dataset_experiment}{The dataset for which the missing values plot will be generated.}

\item{out_dir}{The directory where the plot will be saved.}

\item{out_name}{The filename for the saved plot.}
}
\description{
This function generates a missing values plot using the VIM package's aggr function and saves the plot to a specified directory with a given filename.
}
\examples{
missing_values_plot(dataset_experiment = my_dataset, out_dir = "output", out_name = "missing_plot.png")
}
