% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_factories.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(
  output_name,
  input_name,
  factor_col,
  sample_id_col,
  rowNorm = "NULL",
  transNorm = "NULL",
  scaleNorm = "NULL",
  ref = NULL,
  out_dir
)
}
\arguments{
\item{output_name}{The name of the output target.}

\item{input_name}{The name of the input data.}

\item{factor_col}{The factor column.}

\item{sample_id_col}{The sample ID column.}

\item{rowNorm}{The row normalization method (optional). One of: \code{"QuantileNorm"}, \code{"CompNorm"}, \code{"SumNorm"}, \code{"MedianNorm"}, \code{"SpecNorm"}, or \code{NULL}.}

\item{transNorm}{The transformation normalization method (optional). One of: \code{"LogNorm"}, \code{"CrNorm"}, or \code{NULL}.}

\item{scaleNorm}{The scaling normalization method (optional). One of: \code{"MeanCenter"}, \code{"AutoNorm"}, \code{"ParetoNorm"}, \code{"RangeNorm"}, or \code{NULL}.}

\item{ref}{The reference group for normalization (optional).}

\item{out_dir}{The directory to save plots (optional).}
}
\value{
A list of targets to normalize data.
}
\description{
Normalize data
}
\examples{
normalize(normalized_data, input_data, factor_col = "Group", sample_id_col = "Sample", rowNorm = "CompNorm", transNorm = "LogNorm", scaleNorm = "ParetoNorm", ref = "Creatinine (114.1 / 44.0)", out_dir = "Plots")
}
\seealso{
\code{\link[=normalize_metab]{normalize_metab()}}
}
