% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_csn}
\alias{normalize_csn}
\title{Perform Constant Sum Normalization (CSN)}
\usage{
normalize_csn(dataset_experiment, scaling_factor = 1)
}
\arguments{
\item{dataset_experiment}{A \code{DatasetExperiment} object}

\item{scaling_factor}{Scaling factor for normalization}
}
\value{
Normalized A \code{DatasetExperiment} object
}
\description{
Perform Constant Sum Normalization (CSN)
}
\examples{
normalize_csn(dataset_experiment, scaling_factor = 1)
}
