% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_pqn}
\alias{normalize_pqn}
\title{Perform Probabilistic Quotient normalization (PQN)}
\usage{
normalize_pqn(dataset_experiment, qc_label, factor_name, ...)
}
\arguments{
\item{dataset_experiment}{A \code{DatasetExperiment} object}

\item{qc_label}{Label for quality control samples}

\item{factor_name}{Name of the factor to use for normalization}
}
\value{
Normalized A \code{DatasetExperiment} object
}
\description{
Perform Probabilistic Quotient normalization (PQN)
}
\examples{
normalize_pqn(dataset_experiment, qc_label, factor_name)
}
