% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpr2.R
\name{pcpr2}
\alias{pcpr2}
\title{Run PCPR2 analysis on metabolomic data}
\usage{
pcpr2(
  dataset_experiment,
  variables = colnames(dataset_experiment$sample_meta),
  log_transform = TRUE,
  pct_threshold = 0.8,
  out_dir = NULL,
  out_name = "pcpr2"
)
}
\arguments{
\item{dataset_experiment}{The metabolomic dataset as a SummarizedExperiment object.}

\item{variables}{The variables (covariates) to include in the analysis. Default is all variables in the sample metadata.}

\item{log_transform}{Logical indicating whether to perform log transformation and autoscaling on the data. Default is TRUE.}

\item{pct_threshold}{The threshold for percentage of variability explained by covariates. Default is 0.8.}

\item{out_dir}{The output directory to save the plot. If NULL, the plot will not be saved. Default is NULL.}

\item{out_name}{The name of the output plot file. Default is "pcpr2".}
}
\value{
The plot object showing the variability in metabolomic data explained by covariates.
}
\description{
This function performs PCPR2 analysis on metabolomic data to assess the variability explained by covariates.
}
\examples{
# Load the metabolomic dataset
data("metabolomic_dataset")

# Run PCPR2 analysis
pcpr2_result <- pcpr2(dataset_experiment = metabolomic_dataset)

# Plot the result
pcpr2_result

}
