% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_hotelling_pca}
\alias{plot_hotelling_pca}
\title{Generate a PCA Hotelling's T-squared plot}
\usage{
plot_hotelling_pca(dataset_experiment, nPCs = 5)
}
\arguments{
\item{dataset_experiment}{The dataset for which the plot will be generated.}

\item{nPCs}{The number of principal components to include in the analysis.}
}
\value{
A ggplot object displaying the PCA Hotelling's T-squared plot.
}
\description{
This function generates a PCA Hotelling's T-squared plot showing the principal component scores and confidence ellipses.
}
\examples{
plot_hotelling_pca(dataset_experiment = my_dataset, nPCs = 5)
}
