% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Generate an outliers plot}
\usage{
plot_outliers(dataset_experiment, nPCs = 5, out_dir, out_name)
}
\arguments{
\item{dataset_experiment}{The dataset for which the outliers plot will be generated.}

\item{nPCs}{The number of principal components to include in the analysis.}

\item{out_dir}{The directory where the plot will be saved.}

\item{out_name}{The filename for the saved plot.}
}
\description{
This function generates a plot showing outliers detected using Hotelling's T-squared statistic in PCA.
}
\examples{
plot_outliers(dataset_experiment = my_dataset, nPCs = 5, out_dir = "output", out_name = "outliers_plot.png")
}
