% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{qcrsc_correction}
\alias{qcrsc_correction}
\title{Signal drift and batch correction function}
\usage{
qcrsc_correction(dataset_exp, order_col, batch_col, qc_col, qc_label)
}
\arguments{
\item{dataset_exp}{A DatasetExperiment object with samples and variables.}

\item{order_col}{Column indicating the order of samples.}

\item{batch_col}{Column indicating batch information.}

\item{qc_col}{Column indicating quality control information.}

\item{qc_label}{Label for quality control.}
}
\value{
Corrected DatasetExperiment object.
}
\description{
This function performs signal drift and batch correction on a given DatasetExperiment object using the QC-RSC method.
}
\examples{
#' DE <- metaboPipe::MTBLS79
qcrsc_correction(MTBLS79, "run_order", "Batch", "Type", "QC")
}
