% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample.data.modify}
\alias{sample.data.modify}
\title{Function to modify sample metadata of a DatasetExperiment object}
\usage{
sample.data.modify(dataset_exp, sample_meta)
}
\arguments{
\item{dataset_exp}{A DatasetExperiment object.}

\item{sample_meta}{New sample metadata dataframe.}
}
\value{
A DatasetExperiment object with modified sample metadata.
}
\description{
This function replaces the sample metadata in a SummarizedExperiment object with new metadata.
}
\examples{
sample.data.modify(dataset_exp, sample_meta)
}
