% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_metabo}
\alias{save_metabo}
\title{Function to export MetaboAnalyst data}
\usage{
save_metabo(mSet)
}
\arguments{
\item{mSet}{The MetaboAnalyst data object}
}
\value{
Nothing
}
\description{
Function to export MetaboAnalyst data
}
\examples{
save_metabo(mSet)
}
