% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{sort_by_sample_id}
\alias{sort_by_sample_id}
\title{Sort by sample_id}
\usage{
sort_by_sample_id(df)
}
\arguments{
\item{df}{A dataframe with a sample_id column.}
}
\value{
A data frame sorted by sample_id.
}
\description{
Sort by sample_id
}
\examples{
sort_by_sample_id(data)
}
