% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData.R
\docType{data}
\name{ST000284}
\alias{ST000284}
\title{ST000284 Dataset}
\format{
A \code{DatasetExperiment} object
}
\source{
https://www.metabolomicsworkbench.org/data/DRCCMetadata.php?Mode=Study&StudyID=ST000284&StudyType=MS&ResultType=1
}
\usage{
ST000284
}
\description{
This dataset is used as an example in the package.
}
\keyword{datasets}
