% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{warper_batch_correction}
\alias{warper_batch_correction}
\title{Signal drift and batch correction function}
\usage{
warper_batch_correction(
  dataset_exp,
  method,
  order_col,
  batch_col,
  qc_col,
  qc_label
)
}
\arguments{
\item{dataset_exp}{A DatasetExperiment object with samples and variables.}

\item{method}{The batch correction method to use: ComBat, QCRSC.}

\item{order_col}{Column indicating the order of samples.}

\item{batch_col}{Column indicating batch information.}

\item{qc_col}{Column indicating quality control information.}

\item{qc_label}{Label for quality control.}
}
\value{
Corrected DatasetExperiment object.
}
\description{
This function performs signal drift and batch correction on a given DatasetExperiment object using the specified method.
}
\examples{
DE <- metaboPipe::MTBLS79
warper_batch_correction(MTBLS79, "QCRSC", "run_order", "Batch", "Type", "QC")
}
