% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{warper_createExperiment}
\alias{warper_createExperiment}
\title{Process the dataset to create the DatasetExperiment object}
\usage{
warper_createExperiment(
  dataMatrix,
  sampleMetadata,
  variableMetadata,
  experiment_name = "Name",
  experiment_description = "Description"
)
}
\arguments{
\item{dataMatrix}{A matrix with samples as rows and features as columns.}

\item{sampleMetadata}{A data frame with the sample metadata.}

\item{variableMetadata}{A data frame with the variable metadata.}

\item{experiment_name}{The name for the experiment.}

\item{experiment_description}{The description for the experiment.}
}
\value{
A \code{DatasetExperiment} object.
}
\description{
Process the dataset to create the DatasetExperiment object
}
\examples{
warper_createExperiment(dataMatrix, sampleMetadata, variableMetadata, experiment_name = "Name", experiment_description = "Description")
}
