% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warper_export_data}
\alias{warper_export_data}
\title{Function to export data}
\usage{
warper_export_data(dataset_exp, out_dir, out_name)
}
\arguments{
\item{dataset_exp}{A \code{DatasetExperiment} object}

\item{out_dir}{The output directory}

\item{out_name}{The output name of the files}
}
\value{
Nothing
}
\description{
Function to export data
}
\examples{
warper_export_data(dataset_exp, out_dir, out_name)
}
