% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{zero_to_na}
\alias{zero_to_na}
\title{Make 0 and <0 as \code{NA}}
\usage{
zero_to_na(dataset_exp)
}
\arguments{
\item{dataset_exp}{The \code{DatasetExperiment} object.}
}
\value{
A \code{DatasetExperiment} object with values <= 0 replaced by \code{NA}.
}
\description{
Replace 0 values with \code{NA} in a \code{DatasetExperiment}.
}
\examples{
DE <- metaboPipe::ST000284
modified_dataset <- zero_to_na(DE)
}
