% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_factories.R, R/utils.R
\name{export_data}
\alias{export_data}
\title{Export Data}
\usage{
export_data(dataset_exp, out_dir, out_name)

export_data(dataset_exp, out_dir, out_name)
}
\arguments{
\item{dataset_exp}{A \code{DatasetExperiment} object}

\item{out_dir}{The output directory}

\item{out_name}{The output name of the files}

\item{output_name}{The name of the exported dataset.}

\item{input_name}{The name of the input dataset to be exported.}
}
\value{
A list containing the target for exporting the dataset.

Nothing
}
\description{
Exports a dataset to a specified directory with a given name.
}
\examples{
export_data("exported_dataset.csv", my_dataset, "output_directory/")
export_data(dataset_exp, out_dir, out_name)
}
